[feature] Organized components in files and done passwordSecurity component
