+++ app/src/main/java/com/madteam/sunset/design_system/common/Logos.kt
@@ -1,0 +1,19 @@
+package com.madteam.sunset.design_system.common
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.madteam.sunset.R
+
+@Composable
+fun SunsetLogoImage() {
+    Image(
+        modifier = Modifier.size(width = 214.dp, height = 197.dp),
+        painter = painterResource(id = R.drawable.logo_degrade),
+        contentDescription = "Sunset logo degrade"
+    )
+}
+
