+++ app/src/main/java/com/madteam/sunset/welcome/ui/BottomSheetSignUp.kt
@@ -3,24 +3,16 @@ package com.madteam.sunset.welcome.ui
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Mail
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.madteam.sunset.R
 import com.madteam.sunset.R.string
 import com.madteam.sunset.common.design_system.*
-import com.madteam.sunset.ui.theme.secondarySemiBoldBodyM
+import com.madteam.sunset.design_system.common.*
 
 @Composable
 fun BottomSheetSignUp() {
@@ -45,7 +37,7 @@ fun BottomSheetSignUp() {
 fun SignUpCardContent() {
     var emailValue by remember { mutableStateOf("") }
     var passwordValue by remember { mutableStateOf("") }
-    var nameValue by remember { mutableStateOf("") }
+    var usernameValue by remember { mutableStateOf("") }
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         modifier = Modifier.padding(horizontal = 36.dp)
@@ -59,8 +51,10 @@ fun SignUpCardContent() {
         EmailTextField(emailValue = emailValue, onValueChange = { emailValue = it }, endIcon = { SuccessIcon()} )
         CustomSpacer(size = 16.dp)
         PasswordTextField(passwordValue = passwordValue, onValueChange = { passwordValue = it }, endIcon = {PasswordVisibilityOffIcon()} )
+        CustomSpacer(size = 8.dp)
+        PasswordSecurityIndicator(passwordValue = passwordValue)
         CustomSpacer(size = 16.dp)
-        NameTextField(nameValue = nameValue, onValueChange = {nameValue = it})
+        UsernameTextField(usernameValue = usernameValue, onValueChange = {usernameValue = it})
         CustomSpacer(size = 24.dp)
         SmallButtonDark(onClick = { /*TODO*/ }, text = string.sign_up)
         CustomSpacer(size = 16.dp)
