+++ app/src/main/java/com/madteam/sunset/welcome/ui/WelcomeScreen.kt
@@ -1,38 +1,21 @@
 package com.madteam.sunset.welcome.ui
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Button
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.madteam.sunset.R
-import com.madteam.sunset.R.string
-import com.madteam.sunset.common.design_system.*
-import com.madteam.sunset.ui.theme.primaryBoldDisplayM
-import com.madteam.sunset.ui.theme.secondaryRegularBodyL
-import com.madteam.sunset.ui.theme.secondarySemiBoldHeadLineS
+import com.madteam.sunset.common.design_system.CustomSpacer
+import com.madteam.sunset.common.design_system.MainTitle
+import com.madteam.sunset.common.design_system.SubTitle
+import com.madteam.sunset.design_system.common.EmailButton
+import com.madteam.sunset.design_system.common.FacebookButton
+import com.madteam.sunset.design_system.common.GoogleButton
+import com.madteam.sunset.design_system.common.SunsetLogoImage
 
 @Composable
 fun WelcomeScreen() {
